+++ packages/hw-wallet-sdk/src/index.ts
@@ -2,3 +2,4 @@ export * from './hardware-errors-enums';
 export * from './hardware-error-mappings';
 export * from './hardware-error';
 export * from './types';
+export * from './user-rejection';