+++ packages/keyring-eth-trezor/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Surface Trezor user cancellation and rejection errors as `UserCancelled`/`UserRejected` instead of `ErrorCode.Unknown`, fixing uninformative "Unknown error" messages when users reject signing ([#TODO](https://github.com/MetaMask/accounts/pull/TODO))
+  - Complements [MetaMask/core#8490](https://github.com/MetaMask/core/pull/8490) and fixes [metamask-extension#41184](https://github.com/MetaMask/metamask-extension/issues/41184).
+
 ### Changed
 
 - Bump `@metamask/keyring-api` from `^23.1.0` to `^23.3.0` ([#562](https://github.com/MetaMask/accounts/pull/562), [#569](https://github.com/MetaMask/accounts/pull/569))